backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
pymysql==1.0.3
alembic==1.11.1
sqlalchemy-utils==0.41.0
cryptography
requests
boto3
sentry-sdk==1.26.0
prometheus-client==0.17.0